refactor : 거리 정책 테스트 패키지 이동
